v0.17.008 Tasks are now encrypted whilst they are stored in the GTB datastore, using AES encryption.
GTB now displays a prominent warning message if cookies are disabled, as cookies
are required to store the private RSA key, used to decrypt the encrypted AES key.

For each backup job, GTB generates an RSA private and public key, and an AES key.
The AES key is used by the worker to encrypt the local copy of tasks (after
they have been retrieved from the Google Tasks server).
The AES key is encrypted with an RSA public key, and stored in the job record.
The RSA private key is stored in the user's browser as a cookie.
When the user exports tasks, the RSA private key from the cookie is used to
decrypt the AES key from the job record, and the AES key is then used to
decrypt the stored tasks, which are then returned to the user.